move components and sass files around for refactor
