+++ src/js/components/InspectorSidebar.jsx
