+++ src/js/index.js
@@ -1,6 +1,8 @@
 /* eslint no-unused-expressions: 0 */
 'use strict';
 
+// require('../scss/app.scss');
+
 // Additional requires to polyfill + browserify package.
 require('array.prototype.find');
 require('string.prototype.startswith');
